BeamMexico/gvp__Depletion__c

+++ LiquorWarehouseStage_1_0/SFIn/Tables/gvp__Depletion__c.sql
@@ -31,7 +31,7 @@
 	gvp__BDN_Transfers_Out__c decimal(7,2) null,
 	gvp__BDN_Wholesale_Sales__c decimal(8,2) null,
 	gvp__Depletion_Key__c varchar(95) null,
-	gvp__Month__c DECIMAL(3, 1) null,
+	gvp__Month__c INT null,
 	gvp__Year__c varchar(4) null,
 	gvp__Source_File__c varchar(80) null,
 	gvp__Custom_1__c nvarchar(200) null,
